Create HW2
Explore the Spider Dataset
This github repository from Ben Eyal contains an enriched version of the Spider dataset with additional data which we will explore later in the semester.

Your task in this assignment is to perform exploratory data analysis of the Spider dataset.

Provide your answer in the form of a colab notebook hosted on a github repository.

To prepare your exploration, learn how the questions, schema and SQL queries are encoded in the data.  Based on your understanding of the data format, your data exploration should compute the following type of statistical information:

Number of tables per schema, number of fields per tables.  Show this as histograms (decide on sizes 1-5, 5-10, more than 10 for example, and show number of schema in each bin).
Number of queries per schema, number of questions per query, number of questions per schema (Show the results as a histogram or a table according to the case).
Number of words per question (show as histogram - compute bins for the sizes of the questions, and show how many questions fall in each bin - for example, how many questions with less than 10 words, between 10 and 15, etc)
Number of schema items in SQL query (tables, fields, joins).  For each type, show a histogram.
Depth of the SQL abstract syntax (as it is encoded in the Spider format).
Click https://github.com/beneyal/spider_execution_plans link to open resource.